+++ lib/wifi-wand/command_line_interface/help_system.rb
@@ -10,30 +10,30 @@ module WifiWand
         resource_help = model ? model.resource_manager.open_resources.help_string : "[resources unavailable]"
         
         "
-Command Line Switches:                    [wifi-wand version #{WifiWand::VERSION} at https://github.com/keithrbennett/wifiwand]
-
--o {i,j,k,p,y}            - outputs data in inspect, JSON, pretty JSON, puts, or YAML format when not in shell mode
--p wifi_interface_name    - override automatic detection of interface name with this name
--s                        - run in shell mode
+Command Line Switches     [wifi-wand version #{WifiWand::VERSION} at https://github.com/keithrbennett/wifiwand]
+---------------------
+-o {i,j,k,p,y}            - when not in shell mode, outputs data in the following formats: inspect, JSON, pretty JSON, puts, YAML
+-p wifi_interface_name    - specify wifi interface name (overrides auto-detection)
+-s                        - run in shell mode (interactive pry REPL session)
 -v                        - verbose mode (prints OS commands and their outputs)
 
-Commands:
-
+Commands
+--------
 a[vail_nets]              - array of names of the available networks
-ci                        - connected to Internet (not just wifi on)?
-co[nnect] network-name    - turns wifi on, connects to network-name
-cy[cle]                   - turns wifi off, then on, preserving network selection
+ci                        - state of Internet connectivity, defined as both DNS and TCP working
+co[nnect] network-name    - connects to the specified network-name, turning WiFi on if necessary
+cy[cle]                   - toggles wifi on/off state twice, regardless of starting state
 d[isconnect]              - disconnects from current network, does not turn off wifi
-f[orget] name1 [..name_n] - removes network-name(s) from the preferred networks list
+f[orget] name1 [..name_n] - removes network-name(s) from the preferred (saved) networks list
                             in interactive mode, can be a single array of names, e.g. returned by `pref_nets`
 h[elp]                    - prints this help
-i[nfo]                    - a hash of wifi-related information
+i[nfo]                    - a hash of detailed networking information
 na[meservers]             - nameservers: 'show' or no arg to show, 'clear' to clear,
                             or IP addresses to set, e.g. '9.9.9.9  8.8.8.8'
-ne[twork_name]            - name (SSID) of currently connected network
+ne[twork_name]            - name (SSID) of currently connected wifi network
 on                        - turns wifi on
 of[f]                     - turns wifi off
-pa[ssword] network-name   - password for preferred network-name
+pa[ssword] network-name   - password for preferred network name
 pr[ef_nets]               - preferred (saved) networks
 q[uit]                    - exits this program (interactive shell mode only) (see also 'x')
 ro[pen]                   - open resource (#{resource_help})
@@ -46,7 +46,7 @@ x[it]                     - exits this program (interactive shell mode only) (se
 
 When in interactive shell mode:
   * remember to quote string literals.
-  * for pry commands, use prefix `%`.
+  * for pry commands, use prefix `%`, e.g. `%ls`.
 
 "
       end