+++ lib/wifi-wand/models/base_model.rb
@@ -223,15 +223,9 @@ class BaseModel
     info
   end
 
-  # Turns wifi off and then on, reconnecting to the originally connecting network.
+  # Toggles WiFi on/off state twice; if on, turns off then on; else, turn on then off.
   def cycle_network
-    # TODO: Make this network name saving and restoring conditional on it not having a password.
-    # If the disabled code below is enabled, an error will be raised if a password is required,
-    # even though it is stored.
-    # network_name = connected_network_name
-    wifi_off
-    wifi_on
-    # connect(network_name) if network_name
+    wifi_on? ? (wifi_off; wifi_on) : (wifi_on; wifi_off)
   end